added precision

+++ ivy/functional/backends/paddle/experimental/statistical.py
@@ -389,4 +389,4 @@ def igamma(
         result = paddle.divide(paddle.sum(integral), paddle.exp(paddle.lgamma(ai)))
         results.append(result)
 
-    return paddle.to_tensor(results).reshape(a.shape)
+    return paddle.to_tensor(results, dtype="float32").reshape(a.shape)
